upstream(core): Remove the direct dependency from to iota-core to iota-move-build
