+++ crates/iota-cluster-test/Cargo.toml
@@ -34,6 +34,7 @@ iota-indexer.workspace = true
 iota-json.workspace = true
 iota-json-rpc-types.workspace = true
 iota-keys.workspace = true
+iota-move-build.workspace = true
 iota-sdk.workspace = true
 iota-swarm.workspace = true
 iota-swarm-config.workspace = true
