+++ crates/iota-core/Cargo.toml
@@ -69,7 +69,6 @@ iota-genesis-builder.workspace = true
 iota-json-rpc-types.workspace = true
 iota-macros.workspace = true
 iota-metrics.workspace = true
-iota-move-build.workspace = true
 iota-network.workspace = true
 iota-network-stack.workspace = true
 iota-protocol-config.workspace = true
@@ -104,6 +103,7 @@ tower.workspace = true
 
 # internal dependencies
 iota-move.workspace = true
+iota-move-build.workspace = true
 iota-test-transaction-builder.workspace = true
 
 [target.'cfg(not(target_env = "msvc"))'.dev-dependencies]
